+++ src/app.ts
@@ -11,7 +11,7 @@ import bodyParser from "body-parser";
 import errorHandler from "./middleware/error-handler";
 
 import attendeeRouter from "./services/attendee/attendee-router";
-import adminRouter from "./services/checkin/checkin-router";
+import checkinRouter from "./services/checkin/checkin-router";
 import authRouter from "./services/auth/auth-router";
 import eventsRouter from "./services/events/events-router";
 import notificationsRouter from "./services/notifications/notifications-router";
@@ -38,8 +38,8 @@ app.use("/", bodyParser.json());
 
 // API routes
 app.use("/attendee", databaseMiddleware, attendeeRouter);
-app.use("/admin", databaseMiddleware, adminRouter);
 app.use("/auth", databaseMiddleware, authRouter);
+app.use("/checkin", databaseMiddleware, checkinRouter);
 app.use("/events", databaseMiddleware, eventsRouter);
 app.use("/notifications", databaseMiddleware, notificationsRouter);
 app.use("/registration", databaseMiddleware, registrationRouter);
